PP-2655 Split demoservice controller into multiple files and use lodash on link get - Split demoservice controller into one file per controller - Use `lodash.get` to get pay link from prototype in demoservice and prototype link controllers
