+++ app/controllers/make_a_demo_payment/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+exports.index = require('./index_controller')
+exports.edit = require('./edit_controller')
+exports.confirm = require('./confirm_controller')
